requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
aiosqlite==0.20.0
python-jose[cryptography]==3.4.0
bcrypt>=4.0.0
alembic==1.13.2
pydantic-settings==2.1.0
email-validator>=2.0.0
asyncpg==0.29.0
aiosmtplib==3.0.1
jinja2==3.1.4
python-multipart==0.0.9
slowapi>=0.1.9
